Appy Next.JS project style
