--- stock-web-app/src/App.tsx
@@ -1,29 +1,0 @@
-import './App.scss';
-import Menu from './components/Menu';
-import Main from './Main';
-import { MenuItem } from './types/types';
-
-function App() {
-
-  const menuItems : MenuItem[] = [{
-    label: "MenuItem1",
-    link: "#"
-  },{
-    label: "MenuItem2",
-    link: "#"
-  },{
-    label: "MenuItem3",
-    link: "#"
-  }];
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <Menu items={menuItems}/>
-      </header>
-      <Main />
-    </div>
-  );
-}
-
-export default App;
