+++ stock-web-app/src/components/Menu.tsx
@@ -1,4 +1,3 @@
-import './Menu.scss';
 import { MenuItem } from "../types/types";
 
 function Menu({items} : {items: MenuItem[]}){
