+++ stock-web-app/src/components/StockCard.tsx
@@ -1,4 +1,3 @@
-import './StockCard.scss';
 import { useEffect, useState } from "react";
 import { StockBasicInfo } from "../types/types";
 import Backendservice from "../services/backendservice";
