+++ stock-web-app/src/pages/_document.scss
